requirements.txt
aniso8601==9.0.1
attrs==21.4.0
autopep8==1.6.0
click==8.1.3
Flask==2.1.2
Flask-RESTful==0.3.9
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
packaging==21.3
pluggy==1.0.0
protobuf==4.21.1
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
